refactor: 전반적인 파일 구조 개선
